requirements.txt
openai>=1.12.0
anthropic>=0.18.0
sentence-transformers>=2.3.0
scikit-learn>=1.4.0
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
dataclasses-json>=0.6.0
pydantic>=2.6.0
tqdm>=4.66.0
marshmallow>=3.20.0

